fix: resilience double-push, orchestrator failover-aware sendResponse, throw response handling
- Remove duplicate attempt push in ResilienceLayer.execute() catch block
- Add ctx parameter to orchestrator.sendResponse() for failover threshold check
- Add failover loop check in proxy-handler after orchestrator.handle()
- Add throw/error fallback response in proxy-handler for non-stream cases
- Extract stream proxy to stream-proxy.ts, transport.ts re-exports callStream
- Fix request-tracker.ts ternary indentation (eslint --fix)
